Add audit action to the policy verdict log
[ upstream commit b8f78b33c772e65c8559c43fa3a743381c93df03 ]

The audit mode may overwrite policy related verdict and in this
situation it's unclear from the policy verdict log that this
happened. This commit adds a new audit bit to the policy log's struct
and adds a new policy log action 'audit' based on that flag. Related
policy log calls were updated to contain audit flag.

Signed-off-by: Arthur Evstifeev <aevstifeev@gitlab.com>
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>